Fix: Review의 getElapsedTime 메소드에서 long 타입을 %d로 파싱했던 버그 수정

+++ src/main/java/encore/server/domain/review/entity/Review.java
@@ -105,7 +105,7 @@ public class Review extends BaseTimeEntity {
     }
 
     public String getElapsedTime() {
-        long minutesAgo = ChronoUnit.MINUTES.between(this.getCreatedAt(), LocalDateTime.now());
+        int minutesAgo = (int) ChronoUnit.MINUTES.between(this.getCreatedAt(), LocalDateTime.now());
 
         if (minutesAgo < 1) {
             return "방금 전";
